feat: добавлена возможность получения общих фильмов по id пользователей
