--- src/main/resources/clear.sql
@@ -1,17 +1,0 @@
-drop table "film_directors";
-
-drop table "directors";
-
-drop table "film_genres";
-
-drop table "friendships_sent";
-
-drop table "genres";
-
-drop table "likes";
-
-drop table "films";
-
-drop table "mpa_rating";
-
-drop table "users";
